# config/NMDAOpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API CORONA FINANCIERO - ADMISION",
    "description": "API Corona Financiero Admision",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [{
    "url": "http://localhost:8080/admision/api",
    "description": "Generated server url"
  }],
...
  "paths": {
...
    "/perfilamiento/rol": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "listarRoles",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionRolDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Perfilamiento"],
        "operationId": "actualizarRol",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarRolRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["Perfilamiento"],
        "operationId": "crearRol",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearRolRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/no_recomendados/": {
      "put": {
        "tags": ["No Recomendados"],
        "operationId": "actualizarNoRecomendado",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoRecomendadosRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["No Recomendados"],
        "operationId": "registrarNoRecomendado",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoRecomendadosRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/mensaje-rechazo/": {
      "put": {
        "tags": ["Mensajes de Rechazo"],
        "operationId": "actualizarMensajeRechazo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MensajeRechazoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["Mensajes de Rechazo"],
        "operationId": "crearMensajeRechazo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevoMensajeRechazoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/mfa/verificacion": {
      "post": {
        "tags": ["MFA"],
        "operationId": "verificaValidacionMFA",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFARequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/mfa/validar": {
      "post": {
        "tags": ["MFA"],
        "operationId": "validarSMS",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidarMFARequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/mfa/": {
      "post": {
        "tags": ["MFA"],
        "operationId": "enviarSMS",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFARequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/funcionario": {
      "get": {
        "tags": ["Funcionarios"],
        "operationId": "listarFuncionariosVigentes",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionFuncionarioResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Funcionarios"],
        "operationId": "crearFuncionario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FuncionarioRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/perfilamiento/rol/{rolId}": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "obtenerRol",
        "parameters": [{
          "name": "rolId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RolDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Perfilamiento"],
        "operationId": "eliminarRol",
        "parameters": [{
          "name": "rolId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/perfilamiento/perfil/funcionario/{rutFuncionario}": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "listarPerfilesPorFuncionario",
        "parameters": [{
          "name": "rutFuncionario",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerfilResponseDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/no_recomendados": {
      "get": {
        "tags": ["No Recomendados"],
        "operationId": "listarNoRecomendados",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionNoRecomendadosDTO"
                }
              }
            }
          }
        }
      }
    },
    "/no_recomendados/{rut}": {
      "get": {
        "tags": ["No Recomendados"],
        "operationId": "obtenerNoRecomendable",
        "parameters": [{
          "name": "rut",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NoRecomendadosDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["No Recomendados"],
        "operationId": "eliminarNoRecomendado",
        "parameters": [{
          "name": "rut",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/mensaje-rechazo": {
      "get": {
        "tags": ["Mensajes de Rechazo"],
        "operationId": "listarMensajesRechazo",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionMensajeRechazoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/funcionario/existe/{rutFuncionario}": {
      "get": {
        "tags": ["Funcionarios"],
        "operationId": "existeFuncionario",
        "parameters": [{
          "name": "rutFuncionario",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/mensaje-rechazo/{mensajeRechazoId}": {
      "delete": {
        "tags": ["Mensajes de Rechazo"],
        "operationId": "eliminarMensajeRechazo",
        "parameters": [{
          "name": "mensajeRechazoId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ActualizarRolRequestDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "modulo": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
...
      "NoRecomendadosRequestDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "apellidoPaterno": {
            "type": "string"
          },
          "apellidoMaterno": {
            "type": "string"
          },
          "sucursalOrigen": {
            "type": "integer",
            "format": "int64"
          },
          "motivo": {
            "type": "string"
          }
        }
      },
      "MensajeRechazoDTO": {
        "type": "object",
        "properties": {
          "causalRechazoId": {
            "type": "integer",
            "format": "int64"
          },
          "descripcion": {
            "type": "string"
          },
          "mensajeFuncional": {
            "type": "string"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
...
      "CrearRolRequestDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "modulo": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "FuncionarioRequestDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "nombreCompleto": {
            "type": "string"
          },
          "nombreUsuario": {
            "type": "string"
          },
          "sucursalId": {
            "type": "integer",
            "format": "int64"
          },
          "nombreSucursal": {
            "type": "string"
          },
          "cargoId": {
            "type": "integer",
            "format": "int64"
          },
          "nombreCargo": {
            "type": "string"
          }
        }
      },
      "MFARequestDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          }
        }
      },
      "ValidarMFARequestDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          }
        }
      },
      "NuevoMensajeRechazoDTO": {
        "type": "object",
        "properties": {
          "descripcion": {
            "type": "string"
          },
          "mensajeFuncional": {
            "type": "string"
          }
        }
      },
...
      "SucursalDTO": {
        "type": "object",
        "properties": {
          "codigoSucursal": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "zonaGeografica": {
            "type": "string"
          }
        }
      },
...
      "ListarRolDTO": {
        "type": "object",
        "properties": {
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "PaginacionRolDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListarRolDTO"
            }
          }
        }
      },
      "RolDTO": {
        "type": "object",
        "properties": {
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "moduloId": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
...
      "PerfilResponseDTO": {
        "type": "object",
        "properties": {
          "perfilId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
...
      "NoRecomendadosDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "apellidoPaterno": {
            "type": "string"
          },
          "apellidoMaterno": {
            "type": "string"
          },
          "sucursalOrigen": {
            "$ref": "#/components/schemas/SucursalDTO"
          },
          "motivo": {
            "type": "string"
          }
        }
      },
      "PaginacionNoRecomendadosDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "noRecomendados": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoRecomendadosDTO"
            }
          }
        }
      },
...
      "PaginacionMensajeRechazoDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "mensajes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MensajeRechazoDTO"
            }
          }
        }
      },
...
      "FuncionarioResponseDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "nombreCompleto": {
            "type": "string"
          },
          "nombreUsuario": {
            "type": "string"
          },
          "sucursalId": {
            "type": "integer",
            "format": "int64"
          },
          "nombreSucursal": {
            "type": "string"
          },
          "cargoId": {
            "type": "integer",
            "format": "int64"
          },
          "nombreCargo": {
            "type": "string"
          }
        }
      },
      "PaginacionFuncionarioResponseDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "funcionarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FuncionarioResponseDTO"
            }
          }
        }
      },
...
    }
  }
}